+++ labo1/MatrixReloaded/operators/Add.hpp
@@ -11,7 +11,7 @@
  */
 class Add : public Operator {
 public:
-	int apply(int a, int b) const override;
+	long apply(long a, long b) const override;
 };
 
 #endif // MATRIXRELOADED_ADD_HPP
