+++ labo1/MatrixReloaded/operators/Multiply.cpp
@@ -1,9 +1,5 @@
-//
-// Created by Valentin Kaelin on 03.03.22.
-//
-
 #include "Multiply.hpp"
 
-int Multiply::apply(int a, int b) const {
+long Multiply::apply(long a, long b) const {
 	return a * b;
 }
